cargo2port: Align name output at 20 columns
The unaligned output was not pleasing to the eye, even quite
distracting. However, the output aligned at maxlen would result in large
diffs on port updates when crates are added or removed.

Therefore use a compromise between the two and align the crate names at
the fixed with of 20 spaces.