Add startup message before long read and sort op
The script pauses for a scary few seconds at the start, when it reads and sorts the port list. For a portliest of 750 entries, this takes a bit more time than is comfortable. This change prints a simple startup message before the pause. This should reassure the user that the script is running and not hung or crashed.